cloud_functions/requirements.txt
functions-framework
google-cloud-firestore
protobuf>=3.20.0